initial commit: move login and register to page folder root.
